apps/web/pyproject.toml
[project]
name = "purplesector-api"
version = "0.1.0"
description = "PurpleSector Django backend with Inertia.js (React)"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2,<7",
    "django-cors-headers>=4.0",
    "inertia-django>=1.2",
    "django-vite>=3.0",
    "psycopg2-binary>=2.9",
    "trino>=0.330",
    "pydantic>=2.0",
    "uvicorn[standard]>=0.30",
    "gunicorn>=22.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
ai = [
    "langchain>=0.3",
    "langgraph>=0.4",
    "langchain-openai>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["purplesector"]
